(refactor): share schedule spread offset
